+++ src/test/java/com/bitmart/examples/futures/trade/GetAutoRepayment.java
@@ -1,0 +1,35 @@
+package com.bitmart.examples.futures.trade;
+
+import com.bitmart.api.Call;
+import com.bitmart.api.CloudContext;
+import com.bitmart.api.common.CloudException;
+import com.bitmart.api.common.CloudResponse;
+import com.bitmart.api.key.CloudKey;
+import com.bitmart.api.request.contract.prv.AutoRepaymentRequest;
+import com.bitmart.examples.Example;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class GetAutoRepayment {
+
+    private static final String API_KEY = Example.YOUR_API_KEY;
+    private static final String API_SECRET = Example.YOUR_API_SECRET;
+    private static final String API_MEMO = Example.YOUR_API_MEMO;
+
+    public static void main(String[] args) {
+        Call call = new Call(new CloudContext(Example.FUTURES_HOST, new CloudKey(API_KEY, API_SECRET, API_MEMO)));
+
+        try {
+            final CloudResponse cloudResponse = call.callCloud(new AutoRepaymentRequest()
+                    .setStartTime(1770739200L)
+                    .setEndTime(1771257600L)
+                    .setPage(1)
+                    .setSize(100)
+                    .setFromCoinCode("USDT")
+            );
+            System.out.println(cloudResponse.getResponseContent());
+        } catch (CloudException e) {
+            log.error("Error response: " + e.getMessage());
+        }
+    }
+}