+++ src/test/java/com/bitmart/examples/futures/websocket/Assets.java
@@ -5,6 +5,7 @@ import com.bitmart.api.key.CloudKey;
 import com.bitmart.websocket.ContractWebSocket;
 import com.bitmart.websocket.WebSocketCallBack;
 import com.bitmart.websocket.contract.ActionParam;
+import com.bitmart.examples.Example;
 import com.google.common.collect.ImmutableList;
 import lombok.extern.slf4j.Slf4j;
 
@@ -13,9 +14,9 @@ import static com.bitmart.api.common.GlobalConst.CLOUD_FUTURES_WS_PRIVATE_URL;
 @Slf4j
 public class Assets {
 
-    private static final String API_KEY = "your_api_key";
-    private static final String API_SECRET = "your_secret_key";
-    private static final String API_MEMO = "your_memo";
+    private static final String API_KEY = Example.YOUR_API_KEY;
+    private static final String API_SECRET = Example.YOUR_API_SECRET;
+    private static final String API_MEMO = Example.YOUR_API_MEMO;
 
     public static class ReceiveMessage extends WebSocketCallBack {
         @Override