+++ src/test/java/com/bitmart/examples/spot/market/Ticker.java
@@ -1,5 +1,7 @@
 package com.bitmart.examples.spot.market;
 
+import com.bitmart.examples.Example;
+
 import com.bitmart.api.Call;
 import com.bitmart.api.CloudContext;
 import com.bitmart.api.common.CloudException;
@@ -10,7 +12,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class Ticker {
     public static void main(String[] args) {
-        Call call = new Call(new CloudContext());
+        Call call = new Call(new CloudContext(Example.SPOT_HOST));
 
         try {
             // Get Ticker of a Trading Pair (V3)