+++ src/test/java/com/bitmart/examples/spot/websocket/BalanceChange.java
@@ -6,15 +6,16 @@ import com.bitmart.api.key.CloudKey;
 import com.bitmart.websocket.OpParam;
 import com.bitmart.websocket.WebSocketCallBack;
 import com.bitmart.websocket.WebSocketClient;
+import com.bitmart.examples.Example;
 import com.google.common.collect.ImmutableList;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 public class BalanceChange {
 
-    private static final String API_KEY = "your_api_key";
-    private static final String API_SECRET = "your_secret_key";
-    private static final String API_MEMO = "your_memo";
+    private static final String API_KEY = Example.YOUR_API_KEY;
+    private static final String API_SECRET = Example.YOUR_API_SECRET;
+    private static final String API_MEMO = Example.YOUR_API_MEMO;
 
 
     public static class ReceiveMessage extends WebSocketCallBack {